[APP-1116] Support Xrplevm mainnet
