+++ app/src/main/java/wannabit/io/cosmostaion/chain/BaseChain.kt
@@ -152,6 +152,7 @@ import wannabit.io.cosmostaion.chain.evmClass.ChainStory
 import wannabit.io.cosmostaion.chain.evmClass.ChainStratosEvm
 import wannabit.io.cosmostaion.chain.evmClass.ChainTenetEvm
 import wannabit.io.cosmostaion.chain.evmClass.ChainXplaEvm
+import wannabit.io.cosmostaion.chain.evmClass.ChainXrplEvm
 import wannabit.io.cosmostaion.chain.evmClass.ChainZetaEvm
 import wannabit.io.cosmostaion.chain.fetcher.CosmosFetcher
 import wannabit.io.cosmostaion.chain.fetcher.EvmFetcher
@@ -827,6 +828,7 @@ fun allChains(): MutableList<BaseChain> {
     chains.add(ChainXion())
     chains.add(ChainXplaEvm())
     chains.add(ChainXpla())
+    chains.add(ChainXrplEvm())
     chains.add(ChainZenrock())
     chains.add(ChainZetaEvm())
