+++ app/src/main/java/wannabit/io/cosmostaion/chain/evmClass/ChainXrplEvm.kt
@@ -1,0 +1,27 @@
+package wannabit.io.cosmostaion.chain.evmClass
+
+import android.os.Parcelable
+import com.google.common.collect.ImmutableList
+import kotlinx.parcelize.Parcelize
+import org.bitcoinj.crypto.ChildNumber
+import wannabit.io.cosmostaion.chain.AccountKeyType
+import wannabit.io.cosmostaion.chain.BaseChain
+import wannabit.io.cosmostaion.chain.PubKeyType
+
+@Parcelize
+open class ChainXrplEvm : BaseChain(), Parcelable {
+
+    override var name: String = "XRPL EVM Sidechain"
+    override var tag: String = "xrplevm60"
+    override var apiName: String = "xrplevm"
+
+    override var supportEvm: Boolean = true
+    override var coinSymbol: String = "XRP"
+
+    override var accountKeyType = AccountKeyType(PubKeyType.ETH_KECCAK256, "m/44'/60'/0'/0/X")
+    override var setParentPath: List<ChildNumber> = ImmutableList.of(
+        ChildNumber(44, true), ChildNumber(60, true), ChildNumber.ZERO_HARDENED, ChildNumber.ZERO
+    )
+
+    override var evmRpcURL: String = "https://rpc.evm.xrplevm.mainnet.cosmostation.io"
+}
